refactor cat label sort
